move save_figure_if_needed to plotting._engine_utils
