+++ nilearn/plotting/displays/_slicers.py
@@ -2699,37 +2699,3 @@ def get_slicer(display_mode):
 
     """
     return get_create_display_fun(display_mode, SLICERS)
-
-
-def save_figure_if_needed(fig, output_file):
-    """Save figure if an output file value is given.
-
-    Create output path if required.
-
-    Parameters
-    ----------
-    fig: figure, axes instance
-
-    output_file: str, Path or None
-
-    Returns
-    -------
-    None if ``output_file`` is None, ``fig`` otherwise.
-
-    """
-    if output_file is None:
-        return fig
-
-    output_file = Path(output_file)
-    output_file.parent.mkdir(exist_ok=True, parents=True)
-
-    if not isinstance(fig, plt.Figure):
-        fig = fig.figure
-
-    fig.savefig(output_file)
-    if isinstance(fig, plt.Figure):
-        plt.close(fig)
-    else:
-        fig.close()
-
-    return None
