+++ nilearn/plotting/surface/_matplotlib_backend.py
@@ -13,13 +13,13 @@ from nilearn import DEFAULT_DIVERGING_CMAP
 from nilearn.plotting import cm
 from nilearn.plotting._engine_utils import (
     create_colorbar_for_fig,
+    save_figure_if_needed,
     to_color_strings,
 )
 from nilearn.plotting._utils import (
     DEFAULT_TICK_FORMAT,
 )
 from nilearn.plotting.cm import mix_colormaps
-from nilearn.plotting.displays._slicers import save_figure_if_needed
 from nilearn.plotting.surface._utils import (
     DEFAULT_HEMI,
     check_engine_params,
